// src/app/model/employee.ts
export interface Employee {
    houseRentAllowance: number;
    conveyanceAllowance: number;
    employeeCode: number;
    employeeName: string;
    dateOfBirth: Date;
    gender: string;
    department: string;
    designation: string;
    basicSalary: number;
    dearnessAllowance: number;
    totalSalary: number;
}